Feat: introduce CinemaZ provider
