+++ bazarr/app/get_providers.py
@@ -233,6 +233,10 @@ def get_providers_auth():
             'cookies': settings.avistaz.cookies,
             'user_agent': settings.avistaz.user_agent,
         },
+        'cinemaz': {
+            'cookies': settings.cinemaz.cookies,
+            'user_agent': settings.cinemaz.user_agent,
+        },
         'opensubtitles': {
             'username': settings.opensubtitles.username,
             'password': settings.opensubtitles.password,
